Added bin/build-authors that generates the AUTHORS file.
The AUTHORS file we generate is in the same format at that
used in the Perl distribution, and hence very easy to merge.

The file is generated from all the authors that have made
commits that are in the current branch.  Currently no attempt
is made to check that we're actually on the master branch.